requirements.txt
# SocialVision: Advanced Facial Recognition Search Engine
# Python Dependencies for Academic Research Project
# Developer: Mihretab N. Afework (mtabdevt@gmail.com)

# Core Web Framework
streamlit==1.39.0
gradio==4.44.0

# Computer Vision and Machine Learning
opencv-python-headless==4.10.0.84
face-recognition==1.3.0
numpy==1.26.4
scipy==1.14.1

# Data Processing and Analysis
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Image Processing
Pillow==10.2.0
imageio==2.36.0
scikit-image==0.24.0

# Web Scraping and HTTP
requests==2.32.3
beautifulsoup4==4.12.3
urllib3==2.2.3

# Data Validation and Processing
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12

# API Development
fastapi==0.115.4
uvicorn==0.32.0

# Testing and Quality Assurance
pytest==8.3.3
pytest-cov==6.0.0

# Utilities
tqdm==4.67.1
python-dateutil==2.9.0
pytz==2024.2
click==8.1.7

# Configuration Management
pyyaml==6.0.2
toml==0.10.2

# Logging and Monitoring
loguru==0.7.2

# Version Control Integration
gitpython==3.1.43

# Environment Management
python-decouple==3.8
